deleteFromCart(session check)

+++ shop/models/cart.php
@@ -15,11 +15,13 @@ function addToCart($item_id, $session) {
 }
 
 function deleteFromCart($cart_id, $session) {
-    $result = getOneResult("SELECT cart.id as cart_id, items.item_id as item_id, items.item_title as title, quantity FROM cart, items WHERE $cart_id = cart.id AND cart.item_id = items.item_id AND session_id = '$session'");
+    $result = getOneResult("SELECT cart.id as cart_id, items.item_id as item_id, items.item_title as title, quantity, session_id FROM cart, items WHERE $cart_id = cart.id AND cart.item_id = items.item_id AND session_id = '$session'");
 
-    if ($result['quantity'] > 1) {
-        executeSql("UPDATE `cart` SET quantity = quantity - 1 WHERE id = '$cart_id'");
-    } else {
-        executeSql("DELETE FROM cart WHERE id = '$cart_id'");
+    if ($result["session_id"] == $session) {
+        if ($result['quantity'] > 1) {
+            executeSql("UPDATE `cart` SET quantity = quantity - 1 WHERE id = '$cart_id'");
+        } else {
+            executeSql("DELETE FROM cart WHERE id = '$cart_id'");
+        }
     }
 }
